xds: move balancergroup and weightedtarget our of xds directory
